fix: solc just-in-time installation issue
Between this not resolving and having race conditions, it was not good to rely on forge to auto-download things. Overall hidden install paths are a bit cursed at CI scale.

This downloads it more manually and pins the path in foundry.toml. To avoid rate limiting issues, we cache solc ourselves.