feat: more robust oracle interface change check
The oracle interface change check we had was bad as it didn't check the full signature but only the oracle handler names. In this PR I change it to check the full signature.

The code looks too complicated because we need to manually use the typescript parser to extract the method signatures. Unfortunately there doesn't seem to be a simpler way to do this.