feat!: split private event emission from delivery
The meat of this change is the following:

```diff
- self.emit(event, recipient, delivery_method)
+ self.emit(event).delivery(recipient, delivery_method)
```

This clears lots of things internally, exposing concepts that were hidden inside functions. For end users, it most notably provides a way for the same event to be delivered to multiple recipients, which was impossible before. It also nicely mirrors how notes are handled, which should help these concepts land better.

There's multiple things that are wrong about events right now, both in terms of the code structure and in terms of behavior (e.g. we don't commit to the event type atm), but I wanted to first get this changeset out so that we can include it in the next release and get it in the hands of users.