feat!: single use claims
Adds a state var type: `SingleUseClaim`, which is a useful primitive to build semaphore protocols (see [#1590](https://github.com/AztecProtocol/aztec-packages/issues/15901) for a detailed discussion of semaphores and alternative implementations of this primitive).

It also ports the `PrivateVoting` contract example to rely on it, which shows that the primitive helps abstract contract devs from low-level details.

Closes F-132
Closes https://github.com/AztecProtocol/aztec-packages/issues/15901