feat: Provisional blocks can be added to archiver
This PR makes a number of changes to the archiver.

1. Blocks are now stored separately from checkpoints within the archiver store.
2. Blocks can be added directly to the archiver as well as being synced via checkpoints from L1.
3. Blocks added independently are then deemed to be 'checkpointed' once their checkpoint is received from L1.
4. The archiver now supports multiple blocks per checkpoint.

Some APIs are now considered legacy and will be removed as part of further work. Previously we had `L2Block` and `PublishedL2Block`. We now have `L2BlockNew` and `CheckpointedL2Block`. There are currently APIs for each and this will be updated.

Some operations such manually rolling back blocks still assume 1 block per checkpoint. This will be fixed when we tackle re-orgs of the provisional chain more generally.

Co-authored-by: Santiago Palladino <santiago@aztec-labs.com>