chore: clean up contract verify flow
Contract verification is now enabled only when deploying to mainnet or sepolia AND `ETHERSCAN_API_KEY` env var is set
We error out early if VERIFY_CONTRACTS is set and we would not be able to deploy. but essentially the only purpose VERIFY_CONTRACTS has now is to enable this as a hard error.

Also, we refactor logic in forge deploys to reduce duplication